ai_worker/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
motor>=3.1.1
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.5
pydantic>=2.0.0
python-dotenv>=0.19.0
#tensorflow>=2.0.0
#librosa>=0.9.0
#scipy>=1.7.0
lxml>=4.9.0
#torch>=2.0.0
#numpy==1.21.0
#scikit-learn>=1.0.0
pyjwt